Conflicts:
	README.md
	locscale/__version__.py
	setup.cfg
	setup.py

+++ locscale/utils/file_tools.py
@@ -1,8 +1,6 @@
 
 ## FILE HANDLING FUNCTIONS
-
-from posixpath import isabs
-
+import sys 
 
 def check_dependencies():
     
@@ -160,19 +158,29 @@ def get_input_file_directory(args):
     return input_folder
 
     
-def copy_file_to_folder(full_path_to_file, new_folder):
+def copy_file_to_folder(full_path_to_file, new_folder, mapfile=False):
     import shutil
     import os
     
     source = full_path_to_file
     file_name = os.path.basename(source)
     destination = os.path.join(new_folder, file_name)
-    shutil.copyfile(source, destination)
+    if not os.path.exists(destination):
+        shutil.copyfile(source, destination)
+    else:
+        print(f"File {destination} already exists")
     
-    return destination
+    if mapfile:
+        # Check the axis order 
+        from locscale.preprocessing.headers import check_axis_order
+        xyz_axis_order_path = check_axis_order(destination, use_same_filename=True)
+        return xyz_axis_order_path
+    else:
+        return destination
 
 def change_directory(args, folder_name):
     import os    
+    import locscale
     from locscale.utils.file_tools import copy_file_to_folder
     
     # Get the input folder
@@ -201,22 +209,170 @@ def change_directory(args, folder_name):
     for arg in vars(args):
         value = getattr(args, arg)
         if isinstance(value, str):
-            if os.path.exists(value) and arg != "outfile" and arg != "output_processing_files":
+            if os.path.exists(value) and arg != "outfile" and arg != "output_processing_files" and arg != "emmap_path" and arg != "mask":
                 new_location=copy_file_to_folder(value, new_directory)
                 setattr(args, arg, new_location)
+            elif arg == "emmap_path" or arg == "mask":
+                new_emmap_path = copy_file_to_folder(value, new_directory, mapfile=True)
+                setattr(args, arg, new_emmap_path)
         if isinstance(value, list):
             if arg == "halfmap_paths":
                 halfmap_paths = value
                 halfmap1_path = halfmap_paths[0]
                 halfmap2_path = halfmap_paths[1]
 
-                new_halfmap1_path = copy_file_to_folder(halfmap1_path, new_directory)
-                new_halfmap2_path = copy_file_to_folder(halfmap2_path, new_directory)
+                new_halfmap1_path = copy_file_to_folder(halfmap1_path, new_directory, mapfile=True)
+                new_halfmap2_path = copy_file_to_folder(halfmap2_path, new_directory, mapfile=True)
                 new_halfmap_paths = [new_halfmap1_path,new_halfmap2_path]
                 setattr(args, arg, new_halfmap_paths)
     
+    # Set the logger file path 
+    log_file_path = os.path.join(new_directory, "locscale.log")
+    setattr(args, "logfile_path", log_file_path)
+    logger = setup_logger(log_file_path)
+    logger.info("Starting LocScale program")
+    logger.info("LocScale version: {}".format(locscale.__version__))
+    logger.info("LocScale installed on: {}".format(locscale.__installation_date__))
+    logger.info("-"*80)
+    setattr(args, "logger", logger)
+    setattr(args, "input_folder", input_folder)
     return args
 
+class RedirectStdoutToLogger:
+    # This class was written with the help of chatGPT, model: GPT4
+    def __init__(self, logger, show_progress=True, wait_message="Please wait"):
+        self.logger = logger
+        self._stdout = sys.stdout
+        self.show_progress = show_progress
+        self.symbols = ['/', '-', '\\', '|']
+        self.index = 0
+        self.wait_message = wait_message
+        self.old_stdout = None
+    def __enter__(self):
+        self.old_stdout = sys.stdout
+        self.old_stdout.flush()
+        sys.stdout = self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        sys.stdout = self.old_stdout
+        if self.show_progress:
+            sys.stdout.write('\n')
+            sys.stdout.flush()
+            sys.stdout.write('Done!\n')
+            sys.stdout.flush()
+        if exc_type is not None:
+            self.logger.error(f'Exception: {exc_type}, {exc_val}, {exc_tb}')
+            return False
+        
+        
+    def write(self, content):
+        # Print rotating symbol to stderr if show_progress is True
+        if self.show_progress:
+            self.old_stdout.write(f'\r{self.wait_message}... {self.symbols[self.index % len(self.symbols)]}')
+            self.index += 1
+            self.old_stdout.flush()
+
+        # Avoid logging newline characters
+        if content.rstrip():
+            self.logger.info(content.rstrip())
+
+    def flush(self):
+        pass
+
+
+def run_command_with_filtered_output(command, logger, filter_string=""):
+    import subprocess 
+    print("Running command: {}".format(" ".join(command)))
+    process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True, universal_newlines=True, bufsize=1)
+    
+    for line in process.stdout:
+        # Decode the line
+        line = line.strip()
+        # check if the line contains the filter string
+        logger.info(line)
+        if filter_string in line:
+            # Write the line to the log file
+            print(line)
+    
+    return_code = process.wait()
+    return return_code
+
+def print_downward_arrow(tab_level=0):
+    arrow_string = "\u2193"
+    print("\t"*tab_level + arrow_string)
+    
+    
+def print_ADP_statistics(bfactor_array):
+    import numpy as np
+    print("ADP statistics:")
+    print("Mean: {}".format(np.mean(bfactor_array)))
+    print("Median: {}".format(np.median(bfactor_array)))
+    print("Standard deviation: {}".format(np.std(bfactor_array)))
+    print("Minimum: {}".format(np.min(bfactor_array)))
+    print("Maximum: {}".format(np.max(bfactor_array)))
+    print("")
+    
+class RedirectOutputToLogger:
+    # This class was written with the help of chatGPT, model: GPT4
+    def __init__(self, log_func, show_progress=True, wait_message="Please wait"):
+        self.log_func = log_func
+        self._stdout = sys.stdout
+        self._stderr = sys.stderr
+        self.show_progress = show_progress
+        self.symbols = ['/', '-', '\\', '|']
+        self.index = 0
+        self.wait_message = "Processing"
+
+    def __enter__(self):
+        sys.stdout = self
+        sys.stderr = self
+        if self.show_progress:
+            sys.stderr.write(f'\r{self.wait_message}... {self.symbols[0]}')
+            sys.stderr.flush()
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        sys.stdout = self._stdout
+        sys.stderr = self._stderr
+        if self.show_progress:
+            sys.stderr.write("\n")
+            sys.stderr.flush()
+            sys.stderr.write("Done!           \n")
+            sys.stderr.flush()
+
+    def write(self, content):
+        # Print rotating symbol to stderr if show_progress is True
+        if self.show_progress:
+            sys.stderr.write(f'\r{self.progress_message}... {self.symbols[self.index % len(self.symbols)]}')
+            sys.stderr.flush()
+            self.index += 1
+
+        # Avoid logging newline characters
+        if content.rstrip():
+            self.log_func(content.rstrip())
+
+    def flush(self):
+        pass
+        
+def pretty_print_dictionary(d, indent=1):
+    import numpy as np
+    from textwrap import fill
+    result = ""
+    for key, value in d.items():
+        # Handle large lists, tuples, and NumPy arrays
+        if isinstance(value, (list, tuple)) and len(value) > 10:
+            value_str = f'{type(value).__name__} of length {len(value)}'
+        elif isinstance(value, np.ndarray):
+            value_str = f'numpy array with shape {value.shape}'
+        elif isinstance(value, dict):
+            value_str = '\n' + pretty_print_dictionary(value, indent + 4)
+        else:
+            value_str = str(value)
+        
+        # Wrap the text if it's too long
+        wrapped_value_str = fill(value_str, width=60, subsequent_indent=' ' * (indent + 4)) 
+        result += '\t' * indent + f'{key}: \t {wrapped_value_str}\n'
+    
+    return result
 def generate_filename_from_halfmap_path(in_path):
     ## find filename in the path    
     import os
@@ -263,7 +419,84 @@ def get_emmap_path_from_args(args):
     
     return emmap_path, shift_vector
         
-        ## TBC
+def set_modality_based_on_input(args):
+    '''
+    This function compares all the different ways to get a model map and returns the best option based on the inputs provided
+    Different cases available are: 
+    1) Complete atomic model provided (or the model map is provided)
+    2) Partial atomic model provided
+    3) No atomic model provided 
+    
+    Case 1: Complete atomic model provided
+    - if the input is a coordinate file, then refine the model by default and generate a model map
+    - if the input is a map file, then use the map file as the model map
+    
+    Case 2: Partial atomic model provided
+    - complete the atomic model using pseudo-atomic model building. Then refine the integrated model and generate model map
+    
+    Case 3: No atomic model provided
+    - Predict the model map using a neural network and return the predicted model map
+    - If specified, build a pseudo-atomic model, refine and generate model map   
+    '''
+    modalities = ["full_atomic_model_refine_and_map", "map_input_use_directly", "partial_atomic_model_build_and_refine", "no_model_predict", "no_model_build_and_refine"]
+    
+    ## Check input files
+    emmap_absent = True
+    if args.emmap_path is not None:
+        if is_input_path_valid([args.emmap_path]):
+            emmap_absent = False
+    
+    half_maps_absent = True
+    if args.halfmap_paths is not None:
+        if is_input_path_valid([args.halfmap_paths[0], args.halfmap_paths[1]]):
+            half_maps_absent = False
+    
+    mask_absent = True
+    if args.mask is not None:
+        if is_input_path_valid([args.mask]):
+            mask_absent = False
+    
+    model_map_absent = True
+    if args.model_map is not None:
+        if is_input_path_valid([args.model_map]):
+            model_map_absent = False
+    
+    model_coordinates_absent = True
+    if args.model_coordinates is not None:
+        if is_input_path_valid([args.model_coordinates]):
+            model_coordinates_absent = False
+    
+    emmap_present, half_maps_present = not(emmap_absent), not(half_maps_absent)
+    model_map_present, model_coordinates_present = not(model_map_absent), not(model_coordinates_absent)
+    
+    pseudomodel_not_required = ["map_input_use_directly", "full_model_input_no_refine","full_model_input_refine_and_map", "no_reference","predict_model_map"]
+    pseudomodel_required = ["partial_model_input_build_and_refine", "pseudo_model_build_and_refine"]
+    if model_map_present:
+        modality = "map_input_use_directly"
+    elif model_coordinates_present:
+        if args.complete_model:
+            modality = "partial_model_input_build_and_refine"
+        else:
+            if args.skip_refine:
+                modality = "full_model_input_no_refine"
+            else:
+                modality = "full_model_input_refine_and_map"
+    else:
+        if args.no_reference:
+            modality = "no_reference"
+        else:
+            if args.build_using_pseudomodel:
+                modality = "pseudo_model_build_and_refine"
+            else:
+                modality = "predict_model_map"
+    
+    args.modality = modality
+    args.use_theoretical_profile = modality in pseudomodel_required 
+    args.run_type = "locscale"
+    
+    print("Running LocScale with modality: {}".format(modality))
+    return args
+    
 
 def is_input_path_valid(list_of_test_paths):
     '''
@@ -305,8 +538,37 @@ def simple_test_model_to_map_fit(args):
 
     return correlation
 
+def check_for_refmac(tolerate=False):
+    import os
+    from shutil import which
+    
+    refmac5_path = which("refmac5")
+    
+    if refmac5_path is None:
+        if not tolerate:
+            raise Exception("Refmac5 is not installed. Please install refmac5 and add it to your path")
+        else:
+            print("Refmac5 is not installed. Please install refmac5 and add it to your path")
+    else:
+        print("Refmac5 is installed at {}".format(refmac5_path))
+        print("If you want to use a different binary please use the --refmac5_path option or alias it to refmac5")
 
+def setup_logger(log_path: str):
+    from loguru import logger
+    try:
+        logger.remove(handler_id=0)  # Remove pre-configured sink to sys.stderror
+    except ValueError:
+        pass
 
+    logger.add(
+        log_path,
+        format="{time:YYYY-MM-DD at HH:mm:ss} | {level} | {message}",
+        backtrace=True,
+        enqueue=True,
+        diagnose=True,
+    )
+    return logger
+    
 def check_user_input(args):
     '''
     Check user inputs for errors and conflicts
@@ -354,6 +616,7 @@ def check_user_input(args):
         if is_input_path_valid([args.model_coordinates]):
             model_coordinates_absent = False
     
+    hybrid_locscale = args.complete_model
     ## Rename variables
     emmap_present, half_maps_present = not(emmap_absent), not(half_maps_absent)
     model_map_present, model_coordinates_present = not(model_map_absent), not(model_coordinates_absent)
@@ -370,15 +633,13 @@ def check_user_input(args):
     
     ## If neither model map or model coordinates are provided, then users cannot use --skip_refine flags
     if model_coordinates_absent and model_map_absent:
-        if args.skip_refine:
-            raise UserWarning("You have not provided a Model Map or Model Coordinates. Performing REFMAC refinement is essential for \
-                              succesful operation of the procedue. Please do not raise the --skip_refine flag")
-        
-        if args.ref_resolution is None:
-            raise UserWarning("You have not provided a Model Map or Model Coordinates. To use REFMAC refinement, resolution target is necessary. \
-                              Please provide a target resolution using -res or --ref_resolution")
+        warn_against_skip_refine(args, tolerate=False)
+
+        if args.build_using_pseudomodel:
+            check_and_warn_about_ref_resolution(args)
+
                             
-    if model_coordinates_present:
+    if model_coordinates_present and not hybrid_locscale:
         # Check the model to map fit
         correlation = simple_test_model_to_map_fit(args)
         correlation_threshold = 0.3
@@ -387,20 +648,11 @@ def check_user_input(args):
         
 
     if model_coordinates_present and model_map_absent:
-        if args.skip_refine:
-            print("You have asked to skip REFMAC refinement. Atomic bfactors from the input model will be used for simulating Model Map")
-        else:
-            if args.ref_resolution is None:
-                raise UserWarning("You have provided Model Coordinates. By default, the model bfactors will be refined using REFMAC. \
-                                  For this, a target resolution is required. Please provide this resolution target using -res or --ref_resolution. \
-                                      Instead if you think model bfactors are accurate, then raise the --skip_refine flag to ignore bfactor refinement.")
-            
-
+        warn_against_skip_refine(args, tolerate=True)
+        
     if model_coordinates_present and args.complete_model:
-        if args.skip_refine:
-            raise UserWarning("You have asked to skip REFMAC refinement. \
-                                However, you have asked to complete a partially built model. This requires a refined pseudo-atomic model. \
-                                Please do not raise the --skip_refine flag")
+        warn_against_skip_refine(args, tolerate=False)
+        check_and_warn_about_ref_resolution(args)
         
     ## Check for window size < 10 A
     if args.window_size is not None:
@@ -435,6 +687,24 @@ def check_user_input(args):
         ## Pause 
         import time
         time.sleep(2)
+        
+    ## Find the modalities to generate reference map based on user inputs 
+
+def warn_against_skip_refine(args, tolerate):
+    if args.skip_refine:
+        if not tolerate:
+            raise UserWarning("You have asked to skip REFMAC refinement. \
+                                However, you have asked to complete a partially built model. This requires a refined pseudo-atomic model. \
+                                Please do not raise the --skip_refine flag")
+
+        if tolerate: 
+            print("Warning: You have asked to skip REFMAC refinement. \
+                    Please make sure that the atomic ADPs are refined. LocScale performance maybe severely affected if the ADPs are not refined")
+                
+def check_and_warn_about_ref_resolution(args):
+    if args.ref_resolution is None:
+        raise UserWarning("Please provide a reference resolution using the --ref_resolution flag")
+                
 
 def get_cref_from_inputs(parsed_inputs):
     from locscale.include.emmer.ndimage.filter import get_cosine_mask
